web/api-sr/requirements.txt
audioread==2.1.6
cffi==1.11.5
Click==7.0
decorator==4.3.2
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.1
librosa==0.6.2
llvmlite==0.27.0
MarkupSafe==1.1.0
numba==0.42.0
numpy==1.16.1
pycparser==2.19
PySoundFile==0.9.0.post1
redis==3.1.0
resampy==0.2.1
rq==0.13.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
Werkzeug==0.14.1
